Create settings view for users
What's done:
* Created settings view for users
Closes #646
