+++ save-frontend/src/main/kotlin/org/cqfn/save/frontend/externals/fontawesome/Icons.kt
@@ -37,3 +37,5 @@ external val faArrowRight: dynamic
 external val faLightbulb: dynamic
 external val faCog: dynamic
 external val faHome: dynamic
+external val faEnvelope: dynamic
+external val faKey: dynamic
